refactor(Car.java): 이동 거리 추상화 적용
이동 거리 반환 시, int 값이 아닌 MovingDistance 객체 반환하도록 수정한다.
- MovingDistance.java 인터페이스 생성
- 기존 MovingDistance.java 이름을 CarMovingDistance.java 로 변경
- CarMovingDistance.java, FixedMovingDistance.java 에서 MovingDistance.java 를 구현
