+++ src/main/java/racingcar/car/FixedMovingDistance.java
@@ -1,13 +1,19 @@
 package racingcar.car;
 
-public class FixedMovingDistance extends MovingDistance {
+public class FixedMovingDistance implements MovingDistance {
 	private final int movingDistance;
 
 	public FixedMovingDistance(int movingDistance) {
 		this.movingDistance = movingDistance;
 	}
 
+	@Override
 	public int getDistance() {
 		return movingDistance;
+	}
+
+	@Override
+	public void increase() {
+
 	}
 }
